Bus: Handle CPU-internal narrow writes
The full 32-bits of the GPR are used.

Fixes SoundScope in the BIOS Shell.

+++ src/core/cpu_recompiler_thunks.h
@@ -20,16 +20,16 @@ void CheckAndUpdateICache(u32 pc, u32 line_count);
 u64 ReadMemoryByte(u32 address);
 u64 ReadMemoryHalfWord(u32 address);
 u64 ReadMemoryWord(u32 address);
-u32 WriteMemoryByte(u32 address, u8 value);
-u32 WriteMemoryHalfWord(u32 address, u16 value);
+u32 WriteMemoryByte(u32 address, u32 value);
+u32 WriteMemoryHalfWord(u32 address, u32 value);
 u32 WriteMemoryWord(u32 address, u32 value);
 
 // Unchecked memory access variants. No alignment or bus exceptions.
 u32 UncheckedReadMemoryByte(u32 address);
 u32 UncheckedReadMemoryHalfWord(u32 address);
 u32 UncheckedReadMemoryWord(u32 address);
-void UncheckedWriteMemoryByte(u32 address, u8 value);
-void UncheckedWriteMemoryHalfWord(u32 address, u16 value);
+void UncheckedWriteMemoryByte(u32 address, u32 value);
+void UncheckedWriteMemoryHalfWord(u32 address, u32 value);
 void UncheckedWriteMemoryWord(u32 address, u32 value);
 
 } // namespace Recompiler::Thunks
